We actually already have a UI class in here.
